server/requirements.txt
alembic==1.9.2
bidict==0.22.1
boto3==1.28.24
botocore==1.31.24
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
gevent==23.9.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-engineio==4.8.2
python-socketio==5.11.0
s3transfer==0.6.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.18
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
